match_teacher-main/requirements.txt
numpy
pandas
scipy
scikit-learn
matplotlib
seaborn
plotly
altair
bokeh
pillow
opencv-python
jupyter
requests
google-generativeai
python-dotenv
st-paywall
google_auth
google-auth-oauthlib